Rename Async Runner and introduce serial execution
There is a bug in the "parallel" runner (formerly called Async) by
which there is a crash after new tests are scheduled after the first
round of tests has been scheduled.

For this reason, introduce the simple "serial" runner that relies
on a FIFO queue provided by libdispatch, which guarantees that
basically we are not going to have issues with threads.